requirements.txt
aiofiles==0.8.0; python_version >= "3.6" and python_version < "4.0" and python_full_version >= "3.7.3" and python_full_version < "4.0.0"
anyio==3.6.1; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.7"
apscheduler==3.9.1; python_full_version >= "3.7.3" and python_version < "4" and python_full_version < "4.0.0"
asgiref==3.5.2; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.7"
backports.zoneinfo==0.2.1; python_full_version >= "3.7.3" and python_version < "3.9" and python_full_version < "4.0.0" and python_version >= "3.6"
click==8.1.3; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.7"
colorama==0.4.5; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.7" and sys_platform == "win32" and platform_system == "Windows"
fastapi==0.78.0; python_full_version >= "3.7.3" and python_full_version < "4.0.0"
h11==0.12.0; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.7"
httptools==0.4.0; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.7"
idna==3.3; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.7"
importlib-metadata==4.11.4; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.7" and python_version < "3.8"
loguru==0.6.0; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.5"
msgpack==1.0.4; python_full_version >= "3.7.3" and python_full_version < "4.0.0"
multidict==6.0.2; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.7"
nonebot-adapter-onebot==2.1.1; python_full_version >= "3.7.3" and python_full_version < "4.0.0"
nonebot-adapter-telegram==0.1.0b2; python_full_version >= "3.7.3" and python_full_version < "4.0.0"
nonebot-plugin-apscheduler==0.1.3; python_full_version >= "3.7.3" and python_full_version < "4.0.0"
nonebot2==2.0.0b3; python_full_version >= "3.7.3" and python_full_version < "4.0.0"
pydantic==1.9.1; python_full_version >= "3.7.3" and python_full_version < "4.0.0"
pygtrie==2.4.2; python_full_version >= "3.7.3" and python_full_version < "4.0.0"
python-dotenv==0.20.0
pytz-deprecation-shim==0.1.0.post0; python_full_version >= "3.7.3" and python_version < "4" and python_full_version < "4.0.0" and python_version >= "3.6"
pytz==2022.1; python_full_version >= "3.7.3" and python_version < "4" and python_full_version < "4.0.0"
pyyaml==6.0; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.7"
six==1.16.0; python_full_version >= "3.7.3" and python_version < "4" and python_full_version < "4.0.0"
sniffio==1.2.0; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.7"
starlette==0.19.1; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.6"
tomlkit==0.10.2; python_version >= "3.6" and python_version < "4.0" and python_full_version >= "3.7.3" and python_full_version < "4.0.0"
typing-extensions==4.2.0; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.7" and python_version < "3.8"
tzdata==2022.1; python_full_version >= "3.7.3" and python_version < "4" and python_full_version < "4.0.0" and python_version >= "3.6" and platform_system == "Windows"
tzlocal==4.2; python_full_version >= "3.7.3" and python_version < "4" and python_full_version < "4.0.0" and python_version >= "3.6"
uvicorn==0.17.6; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.7"
uvloop==0.16.0; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.7"
watchgod==0.8.2; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.7"
websockets==10.3; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.7"
win32-setctime==1.1.0; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.5" and sys_platform == "win32"
yarl==1.7.2; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.6"
zipp==3.8.0; python_full_version >= "3.7.3" and python_full_version < "4.0.0" and python_version >= "3.7" and python_version < "3.8"
